moving version to _version
