+++ cmd/bosun/web/web.go
@@ -17,6 +17,7 @@ import (
 	"strings"
 	"time"
 
+	"bosun.org/_version"
 	"bosun.org/cmd/bosun/conf"
 	"bosun.org/cmd/bosun/database"
 	"bosun.org/cmd/bosun/sched"
@@ -26,7 +27,6 @@ import (
 	"bosun.org/opentsdb"
 	"bosun.org/slog"
 	"bosun.org/util"
-	"bosun.org/version"
 
 	"github.com/MiniProfiler/go/miniprofiler"
 	"github.com/gorilla/mux"
